Move demos and change commands
