+++ src/tqdm_publisher/_demos/_demo_command_line_interface.py
@@ -4,18 +4,18 @@ import sys
 import webbrowser
 from pathlib import Path
 
-from tqdm_publisher._demo._multiple._server import run_multiple_bar_demo
-from tqdm_publisher._demo._single_bar._server import run_single_bar_demo
+from tqdm_publisher._demos._multiple_bars._server import run_multiple_bar_demo
+from tqdm_publisher._demos._single_bar._server import run_single_bar_demo
 
 CLIENT_PORT = 1234
 
 DEMOS = {
-    "single_demo": dict(
+    "demo_single": dict(
         subpath="_single_bar",
         server=run_single_bar_demo
     ),
-    "multiple_demo": dict(
-        subpath="_multiple",
+    "demo_multiple": dict(
+        subpath="_multiple_bars",
         server=run_multiple_bar_demo
     )
     # "parallel": "_parallel",
