+++ src/tqdm_publisher/_demos/_multiple_bars/_client.html
@@ -23,7 +23,7 @@
     <header>
         <div>
             <h1>tqdm_progress</h1>
-            <i><small>Create multiple progress bars to test concurrent subscriptions</small></i>
+            <i><small>Click more than once to create multiple progress bars</small></i>
         </div>
 
         <!-- Declare a button to create progress bars -->
