+++ src/tqdm_publisher/_demos/_single_bar/_client.html
